+++ example/lib/pages/home.dart
@@ -10,9 +10,9 @@ class Home extends StatelessWidget {
       className: 'w-full h-full justify-center flex-col items-center bg-gray-200',
       children: [
         WText('Welcome to wind.',
-            className: 'text-primary-500 text-5xl font-bold'),
+            className: 'text-primary-500 text-5xl font-bold text-center'),
         WText('Made with ❤️ by Anılcan Çakır',
-            className: 'text-gray-500'),
+            className: 'text-gray-500 text-center'),
       ],
     );
   }