+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project follows [Semantic Versioning 2.0.0](https://semver.org/spec/v2.0.0.
 ## [0.0.2] - 2025-02-02
 - Added Github workflow for publishing package to pub.dev
 - Updated README.md with installation and usage instructions
+- Added example application to demonstrate usage
 
 ## [0.0.1] - 2025-01-29
 - Initial release